@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+using System.Threading;
 using System.Threading.Tasks;
 using Pinpoint.Core;
 using Pinpoint.Core.Results;
@@ -9,9 +11,9 @@ namespace Pinpoint.Plugin.Bangs
     {
         private const string Description = "Search 10.000+ websites directly via DuckDuckGo bangs.\n\nExamples: \"breaking bad !imdb\", \"diameter of the earth !g\"";
 
-        public PluginMeta Meta { get; set; } = new PluginMeta("DuckDuckGo !Bangs", Description, PluginPriority.Highest);
+        public PluginMeta Meta { get; set; } = new("DuckDuckGo !Bangs", Description, PluginPriority.Highest);
 
-        public PluginSettings UserSettings { get; set; } = new PluginSettings();
+        public PluginSettings UserSettings { get; set; } = new();
 
         public void Unload()
         {
@@ -28,7 +30,7 @@ namespace Pinpoint.Plugin.Bangs
             return query.Parts[0].StartsWith("!") || query.Parts[^1].StartsWith("!");
         }
 
-        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query)
+        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query, [EnumeratorCancellation] CancellationToken ct)
         {
             yield return new BangResult($"https://duckduckgo.com/?q={query.RawQuery.Replace(' ', '+')}");
         }
