@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+using System.Threading;
 using System.Threading.Tasks;
 using Pinpoint.Core;
 using Pinpoint.Core.Results;
@@ -9,9 +11,9 @@ namespace Pinpoint.Plugin.CommandLine
     {
         private const string Description = "Launch CLI programs directly in cmd.\n\nExamples: \">ipconfig\"";
 
-        public PluginMeta Meta { get; set; } = new PluginMeta("Command Line", Description, PluginPriority.Highest);
+        public PluginMeta Meta { get; set; } = new("Command Line", Description, PluginPriority.Highest);
 
-        public PluginSettings UserSettings { get; set; } = new PluginSettings();
+        public PluginSettings UserSettings { get; set; } = new();
 
         public void Unload()
         {
@@ -22,7 +24,7 @@ namespace Pinpoint.Plugin.CommandLine
             return query.Prefix().Equals(">") && query.RawQuery.Length > 1;
         }
 
-        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query)
+        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query, [EnumeratorCancellation] CancellationToken ct)
         {
             yield return new CommandLineResult(query.RawQuery);
         }
