@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Threading;
 using System.Threading.Tasks;
 using Pinpoint.Core.Results;
 using Pinpoint.Core;
@@ -8,11 +10,11 @@ namespace Pinpoint.Plugin.EncodeDecode
 {
     public class EncodeDecodePlugin: IPlugin
     {
-        private readonly string[] _prefixes = {"bin", "hex", "b64"};
+        private static readonly string[] Prefixes = {"bin", "hex", "b64"};
         
-        public PluginMeta Meta { get; set; } = new PluginMeta("Encode/Decode Plugin", PluginPriority.NextHighest);
+        public PluginMeta Meta { get; set; } = new("Encode/Decode Plugin", PluginPriority.NextHighest);
 
-        public PluginSettings UserSettings { get; set; } = new PluginSettings();
+        public PluginSettings UserSettings { get; set; } = new();
 
         public void Unload()
         {
@@ -20,13 +22,13 @@ namespace Pinpoint.Plugin.EncodeDecode
 
         public Task<bool> Activate(Query query)
         {
-            var match = query.RawQuery.Length > 4 && _prefixes.Any(prefix => query.RawQuery.StartsWith(prefix));
+            var match = query.RawQuery.Length > 4 && Prefixes.Any(prefix => query.RawQuery.StartsWith(prefix));
             return Task.FromResult(match);
         }
 
-        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query)
+        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query, [EnumeratorCancellation] CancellationToken ct)
         {
-            var prefix = _prefixes.FirstOrDefault(pre => query.RawQuery.StartsWith(pre));
+            var prefix = Prefixes.FirstOrDefault(pre => query.RawQuery.StartsWith(pre));
 
             var handler = CreateHandler(prefix);
             var queryParts = query.RawQuery.Split(' ');
