@@ -1,10 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net.Http;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using Pinpoint.Core;
 
 namespace Pinpoint.Plugin.Finance
 {
@@ -12,7 +12,7 @@ namespace Pinpoint.Plugin.Finance
     {
         private const string PriceUrl = "https://query1.finance.yahoo.com/v8/finance/chart/{0}?interval=2m&useYfid=true&range=1d";
         private const string SearchUrl = "https://query2.finance.yahoo.com/v1/finance/search?q={0}&quotesCount=1&enableFuzzyQuery=false&newsCount=0";
-        private readonly Dictionary<string, YahooFinanceSearchResponse> _searchCache = new Dictionary<string, YahooFinanceSearchResponse>();
+        private readonly Dictionary<string, YahooFinanceSearchResponse> _searchCache = new();
 
         public YahooFinanceApi(TimeSpan cacheTimeout)
         {
@@ -28,13 +28,9 @@ namespace Pinpoint.Plugin.Finance
                 return _searchCache[ticker];
             }
 
-            var httpResponse = await SendGet(string.Format(SearchUrl, ticker));
-            if (string.IsNullOrEmpty(httpResponse))
-            {
-                return null;
-            }
-
-            var quotes = JObject.Parse(httpResponse)["quotes"].ToArray();
+            var quotes = await HttpHelper.SendGet(string.Format(SearchUrl, ticker),
+                s => JObject.Parse(s)["quotes"].ToArray());
+            
             if (quotes.Length == 0)
             {
                 return null;
@@ -52,13 +48,9 @@ namespace Pinpoint.Plugin.Finance
                 return response;
             }
 
-            var httpResponse = await SendGet(string.Format(PriceUrl, ticker));
-            if (string.IsNullOrEmpty(httpResponse))
-            {
-                return null;
-            }
-
-            var result = JObject.Parse(httpResponse)["chart"]["result"].ToArray();
+            var result = await HttpHelper.SendGet(string.Format(PriceUrl, ticker), 
+                s => JObject.Parse(s)["chart"]["result"].ToArray());
+            
             if (result.Length == 0)
             {
                 return null;
@@ -76,19 +68,6 @@ namespace Pinpoint.Plugin.Finance
             // Cache ticker response
             PriceRepository.Put(ticker, response);
             return response;
-        }
-
-        private async Task<string> SendGet(string url)
-        {
-            try
-            {
-                using var httpClient = new HttpClient();
-                return await httpClient.GetStringAsync(url);
-            }
-            catch (HttpRequestException)
-            {
-                return null;
-            }
         }
     }
 }
