@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Runtime.CompilerServices;
+using System.Threading;
 using System.Threading.Tasks;
 using Pinpoint.Core;
 using Pinpoint.Core.Results;
@@ -9,11 +11,11 @@ namespace Pinpoint.Plugin.HackerNews
     {
         private const string Description = "Browse the frontpage of Hacker News.\n\nExamples: \"hackernews\", \"hacker news\", \"hnews\"";
 
-        private readonly HackerNewsApi _hackerNewsApi = new HackerNewsApi();
+        private readonly HackerNewsApi _hackerNewsApi = new();
 
-        public PluginMeta Meta { get; set; } = new PluginMeta("Hacker News Browser", Description, PluginPriority.Highest);
+        public PluginMeta Meta { get; set; } = new("Hacker News Browser", Description, PluginPriority.Highest);
 
-        public PluginSettings UserSettings { get; set; } = new PluginSettings();
+        public PluginSettings UserSettings { get; set; } = new();
 
         public void Unload()
         {
@@ -25,7 +27,7 @@ namespace Pinpoint.Plugin.HackerNews
             return raw.Length >= 10 && (raw.Equals("hackernews") || raw.Equals("hacker news") || raw.Equals("hnews"));
         }
 
-        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query)
+        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query, [EnumeratorCancellation] CancellationToken ct)
         {
             var stories = await _hackerNewsApi.TopSubmissions();
             for (var i = 0; i < stories.Count; i++)
