@@ -4,6 +4,6 @@ namespace Pinpoint.Plugin.Reddit.Models
 {
     public class RedditApiResultData
     {
-        public List<RedditPostResult> Children { get; set; } = new List<RedditPostResult>();
+        public List<RedditPostResult> Children { get; set; } = new();
     }
 }
