@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net;
 using System.Net.Http;
 using System.Text;
 using System.Threading.Tasks;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using Pinpoint.Core;
 using PinPoint.Plugin.Spotify;
 
@@ -15,8 +13,9 @@ namespace Pinpoint.Plugin.Spotify.Client
     public class SpotifyClient: IDisposable
     {
         private static SpotifyClient _instance;
-        private  readonly HttpClient _spotifyHttpClient = new HttpClient();
+        private static readonly HttpClient SpotifyHttpClient = new();
         private string _accessToken;
+        private const string SpotifyApiBaseUrl = "https://api.spotify.com/v1";
 
         public static SpotifyClient GetInstance()
         {
@@ -36,7 +35,7 @@ namespace Pinpoint.Plugin.Spotify.Client
         public async Task<List<SpotifyResultEntity>> Search(string query, string type)
         {
             var message = CreateRequestMessage(HttpMethod.Get,
-                $"https://api.spotify.com/v1/search?q={query}&type={type}&limit=5");
+                $"{SpotifyApiBaseUrl}/search?q={query}&type={type}&limit=5");
 
             var response = await SendWithRetry(message);
 
@@ -72,10 +71,22 @@ namespace Pinpoint.Plugin.Spotify.Client
                     : new PlayRequest {ContextUri = uri};
             }
 
-            var message = CreateRequestMessage(HttpMethod.Put, "https://api.spotify.com/v1/me/player/play",
+            var message = CreateRequestMessage(HttpMethod.Put, $"{SpotifyApiBaseUrl}/me/player/play",
                 uri != null ? request : null);
 
-            var _ = await SendWithRetry(message);
+            var response = await SendWithRetry(message);
+
+            if (!response.IsSuccessStatusCode)
+            {
+                var availableDevices = await GetAvailableDevices();
+
+                var deviceToPlayOn = availableDevices.FirstOrDefault();
+                if (deviceToPlayOn == null) return;
+
+                var newMessage = CreateRequestMessage(HttpMethod.Put,
+                    $"{SpotifyApiBaseUrl}/me/player/play?device_id={deviceToPlayOn.Id}");
+                await SpotifyHttpClient.SendAsync(newMessage);
+            }
         }
 
         public async Task PlayPauseCurrentTrack()
@@ -84,7 +95,7 @@ namespace Pinpoint.Plugin.Spotify.Client
 
             if (isPlaying)
             {
-                var message = CreateRequestMessage(HttpMethod.Put, "https://api.spotify.com/v1/me/player/pause");
+                var message = CreateRequestMessage(HttpMethod.Put, $"{SpotifyApiBaseUrl}/me/player/pause");
 
                 var _ = await SendWithRetry(message);
             }
@@ -97,31 +108,45 @@ namespace Pinpoint.Plugin.Spotify.Client
         public async Task QueueItem(string uri)
         {
             var message =
-                CreateRequestMessage(HttpMethod.Post, $"https://api.spotify.com/v1/me/player/queue?uri={uri}");
+                CreateRequestMessage(HttpMethod.Post, $"{SpotifyApiBaseUrl}/me/player/queue?uri={uri}");
             var _ = await SendWithRetry(message);
         }
 
         public async Task NextTrack()
         {
-            var message = CreateRequestMessage(HttpMethod.Post, "https://api.spotify.com/v1/me/player/next");
+            var message = CreateRequestMessage(HttpMethod.Post, $"{SpotifyApiBaseUrl}/me/player/next");
             var _ = await SendWithRetry(message);
         }
 
         public async Task PreviousTrack()
         {
-            var message = CreateRequestMessage(HttpMethod.Post, "https://api.spotify.com/v1/me/player/previous");
+            var message = CreateRequestMessage(HttpMethod.Post, $"{SpotifyApiBaseUrl}/me/player/previous");
             var _ = await SendWithRetry(message);
+        }
+
+        private async Task<List<PlayerDevice>> GetAvailableDevices()
+        {
+            var message = CreateRequestMessage(HttpMethod.Get, $"{SpotifyApiBaseUrl}/me/player/devices");
+            var response = await SendWithRetry(message);
+
+            if (!response.IsSuccessStatusCode) return new List<PlayerDevice>();
+
+            var responseJson = await response.Content.ReadAsStringAsync();
+
+            var result = JsonConvert.DeserializeObject<AvailableDevicesResult>(responseJson);
+
+            return result.Devices;
         }
 
         private async Task<bool> SpotifyCurrentlyPlaying()
         {
-            var message = CreateRequestMessage(HttpMethod.Get, "https://api.spotify.com/v1/me/player");
+            var message = CreateRequestMessage(HttpMethod.Get, $"{SpotifyApiBaseUrl}/me/player");
             var response = await SendWithRetry(message);
 
             var responseContent = await response.Content.ReadAsStringAsync();
             var result = JsonConvert.DeserializeObject<PlaybackInfoResult>(responseContent);
 
-            return result.IsPlaying;
+            return result != null && result.IsPlaying;
         }
 
         private async Task ExchangeRefreshToken()
@@ -136,7 +161,7 @@ namespace Pinpoint.Plugin.Spotify.Client
             });
             var request = new HttpRequestMessage(HttpMethod.Post, "https://accounts.spotify.com/api/token") { Content = content };
 
-            var response = await _spotifyHttpClient.SendAsync(request);
+            var response = await SpotifyHttpClient.SendAsync(request);
             if(!response.IsSuccessStatusCode)
             {
                 return;
@@ -157,7 +182,7 @@ namespace Pinpoint.Plugin.Spotify.Client
 
         public void Dispose()
         {
-            _spotifyHttpClient.Dispose();
+            SpotifyHttpClient.Dispose();
         }
 
         private HttpRequestMessage CreateRequestMessage(HttpMethod httpMethod, string requestUri, object requestContent = null)
@@ -169,16 +194,20 @@ namespace Pinpoint.Plugin.Spotify.Client
 
         private async Task<HttpResponseMessage> SendWithRetry(HttpRequestMessage message)
         {
-            var response = await _spotifyHttpClient.SendAsync(message);
+            var response = await SpotifyHttpClient.SendAsync(message);
 
             if (response.IsSuccessStatusCode) return response;
 
             await ExchangeRefreshToken();
 
             var retryMessage = new HttpRequestMessage(message.Method, message.RequestUri)
-            { Headers = { { "Authorization", $"Bearer {_accessToken}" }, { "Accept", "application/json" } }, Content = message.Method != HttpMethod.Get ? message.Content : null};
+            {
+                Headers = {
+                    { "Authorization", $"Bearer {_accessToken}" },
+                    { "Accept", "application/json" }
+            }, Content = message.Method != HttpMethod.Get ? message.Content : null};
 
-            return await _spotifyHttpClient.SendAsync(retryMessage);
+            return await SpotifyHttpClient.SendAsync(retryMessage);
 
         }
     }
@@ -187,5 +216,17 @@ namespace Pinpoint.Plugin.Spotify.Client
     {
         [JsonProperty("is_playing")]
         public bool IsPlaying { get; set; }
+    }
+
+    public class AvailableDevicesResult
+    {
+        [JsonProperty("devices")]
+        public List<PlayerDevice> Devices { get; set; }
+    }
+
+    public class PlayerDevice
+    {
+        [JsonProperty("id")]
+        public string Id { get; set; }
     }
 }
