@@ -1,19 +1,31 @@
-﻿using System.Threading.Tasks;
-using FontAwesome5;
+﻿using System;
+using System.Drawing;
 using Pinpoint.Core.Results;
 using System.Windows.Forms;
-using System.Windows.Input;
-using Pinpoint.Core;
+using Pinpoint.Core.Clipboard;
 
 namespace Pinpoint.Plugin.ClipboardManager
 {
-    public class ClipboardResult : AbstractFontAwesomeQueryResult
+    public class ClipboardResult : AbstractQueryResult
     {
         private readonly IClipboardEntry _clipboardEntry;
 
-        public ClipboardResult(IClipboardEntry clipboardEntry) : base(clipboardEntry.Title, "Select to paste")
+        public ClipboardResult(IClipboardEntry clipboardEntry) : base(clipboardEntry.Title)
         {
             _clipboardEntry = clipboardEntry;
+            Subtitle = $"Copied {MakeReadableTimestamp()}";
+        }
+
+        private string MakeReadableTimestamp()
+        {
+            var today = DateTime.Today;
+            var actual = _clipboardEntry.Copied;
+            if (today.DayOfYear == actual.DayOfYear)
+            {
+                return $"today on {actual:HH:mm}";
+            }
+
+            return actual.ToString("ddd, dd MMM yyy HH':'mm':'ss");
         }
 
         public override void OnSelect()
@@ -32,6 +44,6 @@ namespace Pinpoint.Plugin.ClipboardManager
             ClipboardHelper.PasteClipboard();
         }
 
-        public override EFontAwesomeIcon FontAwesomeIcon => EFontAwesomeIcon.Regular_Copy;
+        public override Bitmap Icon => _clipboardEntry.Icon;
     }
 }
