@@ -1,4 +1,5 @@
 ﻿using System.Diagnostics;
+using System.Drawing;
 using System.IO;
 using FontAwesome5;
 using Pinpoint.Core.Results;
@@ -6,17 +7,16 @@ using Pinpoint.Plugin.Everything.API;
 
 namespace Pinpoint.Plugin.Everything
 {
-    public class EverythingResult : AbstractFontAwesomeQueryResult
+    public class EverythingResult : AbstractQueryResult
     {
         private readonly QueryResultItem _item;
 
-        public EverythingResult(QueryResultItem item, EFontAwesomeIcon icon) : base(
+        public EverythingResult(QueryResultItem item) : base(
             item.ResultType != ResultType.Directory ? item.Name : Path.GetFileName(item.FullPath),
             item.FullPath
             )
         {
             _item = item;
-            FontAwesomeIcon = icon;
             
             Options.Add(new RunAsAdminOption(item.FullPath));
 
@@ -26,11 +26,18 @@ namespace Pinpoint.Plugin.Everything
             }
         }
 
-        public override EFontAwesomeIcon FontAwesomeIcon { get; }
+        public override Bitmap Icon => _item.Icon;
+        
 
         public override void OnSelect()
         {
             Process.Start("explorer.exe", _item.FullPath);
+        }
+
+        public override bool OnPrimaryOptionSelect()
+        {
+            Process.Start("explorer.exe", Path.GetDirectoryName(_item.FullPath));
+            return true;
         }
     }
 }
