@@ -1,6 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Threading;
 using System.Threading.Tasks;
 using Pinpoint.Core;
 using Pinpoint.Core.Results;
@@ -16,11 +18,7 @@ namespace Pinpoint.Plugin.Finance
         public PluginMeta Meta { get; set; } = new PluginMeta("Finance Plugin", Description, PluginPriority.Highest);
 
         public PluginSettings UserSettings { get; set; } = new PluginSettings();
-
-        public void Unload()
-        {
-        }
-
+        
         public async Task<bool> Activate(Query query)
         {
             if (query.Parts.Length != 1 || !query.Prefix().Equals("$") || query.RawQuery.Length < 3)
@@ -32,10 +30,10 @@ namespace Pinpoint.Plugin.Finance
             return ticker.All(ch => char.IsLetter(ch) || ch == '.' || ch == '^') && ticker.Count(char.IsLetter) >= 2;
         }
 
-        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query)
+        public async IAsyncEnumerable<AbstractQueryResult> Process(Query query, [EnumeratorCancellation] CancellationToken ct)
         {
             // $GME => GME
-            var ticker = query.Parts[0].Substring(1);
+            var ticker = query.Parts[0][1..];
             var priceResponseTask = _yahooFinanceApi.LookupPrice(ticker);
             var searchResponseTask = _yahooFinanceApi.Search(ticker);
 
