@@ -1,10 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Configuration;
-using System.Data;
-using System.Linq;
-using System.Threading.Tasks;
 using System.Windows;
+using Microsoft.Extensions.DependencyInjection;
+using Pinpoint.Win.ViewModels;
+using Pinpoint.Win.Views;
 
 namespace Pinpoint.Win
 {
@@ -13,5 +11,35 @@ namespace Pinpoint.Win
     /// </summary>
     public partial class App : Application
     {
+        public new static App Current => (App)Application.Current;
+
+        public IServiceProvider Services { get; } = ConfigureServices();
+
+        private static IServiceProvider ConfigureServices()
+        {
+            var services = new ServiceCollection();
+            services.AddSingleton<MainWindow>();
+            services.AddSingleton<SettingsWindow>();
+            services.AddSingleton<MainViewModel>();
+            services.AddSingleton<SettingsViewModel>();
+            services.AddTransient<PluginTabItem>();
+            services.AddTransient<PluginTabItemViewModel>();
+            return services.BuildServiceProvider();
+        }
+
+        public new MainWindow MainWindow => Services.GetService<MainWindow>();
+
+        public SettingsWindow SettingsWindow => Services.GetService<SettingsWindow>();
+
+        public MainViewModel MainViewModel => Services.GetService<MainViewModel>();
+
+        public SettingsViewModel SettingsViewModel => Services.GetService<SettingsViewModel>();
+
+        public PluginTabItemViewModel PluginTabItemViewModel => Services.GetService<PluginTabItemViewModel>();
+
+        private void App_OnStartup(object sender, StartupEventArgs e)
+        {
+            Services.GetService<MainWindow>()?.Show();
+        }
     }
 }
